pyproject.toml
[project]
name = "y-rig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
    "maya-stubs>=0.4.2",
    "numpy>=2.4.1",
    "ruff>=0.14.14",
]

[tool.pyright]
extraPaths = [
    "./.stubs/",
    "/usr/autodesk/ApplicationPlugins/ngskintools2/Contents/scripts",
    "third_party/mgear/release/scripts/",
    "src"
]
typeCheckingMode = "standard"
reportMissingModuleSource = "information"

[tool.ruff]
line-length = 100
